requirements.txt
cycler==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
Pillow==8.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
simplejson==3.17.5
six==1.16.0
